backend/pyproject.toml
[tool.poetry]
name = "ads-chair-backend"
version = "0.1.0"
description = ""
authors = ["Egor Kuziakov <egor.kuziakov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
python-multipart = "^0.0.9"
prometheus-client = "^0.20.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
